requirements.txt
python-dotenv
langchain
langchain-community
chromadb
tiktoken
duckduckgo-search
beautifulsoup4
openai
langchain-openai